Add an AC_SEARCH_LIBS line for libcurl
Also adds an error message if libcurl is missing.  NB: This shouldn't be a fatal
error; instead, this should ideally just remove -DTADSNET from the CC flags, and
remove the network-related .cpp files from the source list, to build the
non-networked interpreter.  I didn't want to mess with the automake
configuration to that extent, though, so I took the simple way out for now.
This should still show a warning message in case the user intended to build the
network version, but it should be possible to continue the build with the
network functionality disabled.