Don't allocate mutex on the heap using 'new' but make it a normal variable
This crashes when building the debug version of TADS 3. The memory allocator
(t3malloc()) wants to lock the mem_mutex to protect its debugging block list,
but mem_mutex wants to allocate memory via t3malloc() when it does the
'h = new pthread_mutex_t()' to create the pthreads mutex's memory. So the static
initializer for mem_mutex calls OS_Mutex::OS_Mutex(), which calls t3malloc(),
which calls mem_mutex.lock(), which tries to dereference 'h', which isn't
initialized yet since we got here trying to initialize it!

I'm not sure now why I bothered doing that in the first place - it looks to me
like the pthreads definition is specifically designed for allocation as an
ordinary variable.