Disable the Nagle algorithm
Disable the Nagle algorithm on incoming connections to minimize transmit
latency.  We send a lot of very small packets and the client must
acknowledge each one in sequence before the next one can be sent.  It is
normally useful to combine many small packets into a few large ones, but
here it only slows us down.