Fix relative path handling of log input and replay files
When a relative path is given to the --log-input and --replay options,
they are resolved only after we chdir to the game directory and thus
ends up being relative to that. Which is clearly not what anyone would
expect.

Fix the issue by simply converting them to an absolute path before we
chdir.